Move meta-annotation support code out into its own class
